Update blackboard infrastructure significantly
Significantly updates the Blackboard interface and supporting
implementation in order to provide blackboard clients the ability to
observe subsets of the entire blackboard data. This feature is important
as it allows the blackboard to, on the event of an update, notify only
those client components that must react to and handle the update.