Fix GameLogicOrientation threadhold updating
Fixes the GameLogicOrientation class calculateOrientation method so that
it correctly updates the user location to the blackboard when said
location changes by a threshold amount since the last blackboard update.
Previously, the user location would only be updated to the blackboard
when said location changed by a threshold since the last sensor reading;
effectively, the user location was only updated when the orientation
changed rapidly enough to outpace the sensor sampling rate.